Fix to have mvn site succeed.
mvn test will still fail, which is to be expected as I have not fixed any of the Javadoc errors. Have the site report them in an easy to parse manner(read website), is much easier than reading from the command line or command line output piped to a file.